alexander/requirements.txt
dropbox
pdfplumber
supabase
PyMuPDF
google-genai
pdf2image
pillow
python-dotenv
